Update README.md
Shangri-La Frontier is the most enjoyable anime I have watched. It is the absolute peak anime about VR Games. I wish I could also play Shangri-La Frontier irl but since I can't I must content myself  by watching Sunraku 😭.

Freiren: Beyond Journey's End is also really enjoyable. It's about what would happen after the heroes defeated the demon lord. It has an interesting premise and it is well executed. The best part of this anime is its magic system and world building. Absolute peak!

Dandadan is about a guy who believes in aliens but doesn't believe into ghost and a girl who believes in ghosts but doesn't believe in aliens. This anime is literally a me and who type of anime. It's not horror despite it about being ghost and its more of an action beat up type of anime. Basically they fight ghost and aliens. I really like the opening song of this one and the best part of the anime is the well written characters.